シェーダー単色表示機能完成、シーン機能実装、物理挙動実装

+++ InvaderGame/GameObject.h
@@ -32,7 +32,7 @@ public:
 		{
 			return nullptr;
 		}
-		int index = std::distance(m_componentVector.begin(), iterator);
+		__int64 index = std::distance(m_componentVector.begin(), iterator);
 		T* component = dynamic_cast<T*>(m_componentVector[index]); // dynamic_castは抽象クラスにしかダウンキャストととして利用できない
 		return component;
 	}
